+++ clients/derivatives_trading_portfolio_margin/src/binance_sdk_derivatives_trading_portfolio_margin/rest_api/models/cancel_margin_account_all_open_orders_on_a_symbol_response_inner_order_reports_inner.py
@@ -55,9 +55,7 @@ class CancelMarginAccountAllOpenOrdersOnASymbolResponseInnerOrderReportsInner(
     cummulative_quote_qty: Optional[StrictStr] = Field(
         default=None, description="Cummulative Quote Qty.", alias="cummulativeQuoteQty"
     )
-    status: Optional[StrictStr] = Field(
-        default=None, description="Enum：completed，processing"
-    )
+    status: Optional[StrictStr] = Field(default=None, description="Status.")
     time_in_force: Optional[StrictStr] = Field(
         default=None, description="Time In Force.", alias="timeInForce"
     )