+++ clients/derivatives_trading_portfolio_margin/src/binance_sdk_derivatives_trading_portfolio_margin/rest_api/models/cancel_margin_account_order_response.py
@@ -50,9 +50,7 @@ class CancelMarginAccountOrderResponse(BaseModel):
     cummulative_quote_qty: Optional[StrictStr] = Field(
         default=None, description="Cummulative Quote Qty.", alias="cummulativeQuoteQty"
     )
-    status: Optional[StrictStr] = Field(
-        default=None, description="Enum：completed，processing"
-    )
+    status: Optional[StrictStr] = Field(default=None, description="Status.")
     time_in_force: Optional[StrictStr] = Field(
         default=None, description="Time In Force.", alias="timeInForce"
     )