+++ clients/derivatives_trading_portfolio_margin/src/binance_sdk_derivatives_trading_portfolio_margin/rest_api/models/query_current_um_open_order_response.py
@@ -59,9 +59,7 @@ class QueryCurrentUmOpenOrderResponse(BaseModel):
         description="BOTH means that it is the position of One-way Mode",
         alias="positionSide",
     )
-    status: Optional[StrictStr] = Field(
-        default=None, description="Enum：completed，processing"
-    )
+    status: Optional[StrictStr] = Field(default=None, description="Status.")
     symbol: Optional[StrictStr] = Field(
         default=None, description="Trade symbol, if existing."
     )