+++ clients/derivatives_trading_portfolio_margin/src/binance_sdk_derivatives_trading_portfolio_margin/websocket_streams/models/algo_order_update_ao.py
@@ -71,6 +71,10 @@ class AlgoOrderUpdateAo(BaseModel):
     rm: Optional[StrictStr] = Field(
         default=None, description="Algo order failed reason"
     )
+    ia: Optional[StrictBool] = Field(
+        default=None,
+        description="Is activated or not, only meaningful for trailing order",
+    )
     additional_properties: Dict[str, Any] = {}
     __properties: ClassVar[List[str]] = [
         "caid",
@@ -98,6 +102,7 @@ class AlgoOrderUpdateAo(BaseModel):
         "tt",
         "gtd",
         "rm",
+        "ia",
     ]
 
     model_config = ConfigDict(
@@ -185,6 +190,7 @@ class AlgoOrderUpdateAo(BaseModel):
                 "tt": obj.get("tt"),
                 "gtd": obj.get("gtd"),
                 "rm": obj.get("rm"),
+                "ia": obj.get("ia"),
             }
         )
         # store additional fields in additional_properties